+++ apps/backend/src/libs/modules/socket/libs/types/types.ts
@@ -1,6 +1,7 @@
 export { type ServerToClientEvents } from "./server-to-client-events.type.js";
 export { type SocketService } from "./socket-service.types.js";
 export {
+	type MeetingAudioSaveDto,
 	type MeetingSummaryActionItemsResponseDto,
 	type MeetingTranscriptionRequestDto,
 	type ValueOf,