+++ apps/bot/src/libs/modules/audio-recorder/libs/enums/enums.ts
@@ -1,1 +1,4 @@
+export { AudioFileType } from "./audio-file-type.enum.js";
 export { AudioRecorderEvent } from "./audio-recorder-event.enum.js";
+export { WatcherEvent } from "./watcher-event.enum.js";
+export { ContentType } from "@meetlytic/shared";